requirements.txt
flask
flask-cors
supabase
huggingface_hub
joblib
numpy
pandas
scikit-learn
python-dotenv
gunicorn
